test: consolidate duplicated decompose-test fixture builders into conftest.py
_fit_response/_split_response/_make_caps were near-identically defined
across test_recursive_decompose.py, test_recursive_decompose_schedule.py,
and test_decompose_budget_share.py; extract one shared copy of each into
tests/conftest.py and import from there.
